refactor: extract per-type rdata codec into DNSpacket.RData (#111)
Move the create_rdata/parse_rdata clauses and their codec-only helpers
(character-strings, NSEC type bitmap, SVCB/HTTPS service params, A/AAAA
fast paths) into DNSpacket.RData, pairing each type's encode and decode
clauses in the same order so they read side by side.

DNSpacket keeps stable @doc false entry points as defdelegates
(create_rdata/3, parse_rdata/4, create_type_bitmap/1, parse_type_bitmap/1,
create_svc_params/1, parse_svc_params/1) since tests call them directly.
The shared name codec (create_domain_name/1, create_character_string/1,
parse_name/3) stays in DNSpacket — it is also on the question/record
framing hot path — and parse_name/3 becomes @doc false public so RData
can decode embedded names with it.

dns_packet.ex drops from 1262 to 705 lines. Behavior is unchanged,
pinned by the round-trip contract suite and the property tests (#107).
